fix: reduce complexity + expand console error
