+++ apps/artemis-web/components/ongoing-hijack-table/ongoing-hijack-table.tsx
@@ -36,113 +36,116 @@ const expandRow: ExpandRowProps<any, number> = {
   renderer: (row) => {
     return (
       <table>
-        <tr>
-          <td>
-            <b>
-              <div data-tip data-for={'hprefix_exp'}>
-                {<span>{'Hijacked Prefix'}</span>}
-              </div>
-              <ReactTooltip html={true} id={'hprefix_exp'}>
-                {'The IPv4/IPv6 prefix that was hijacked.'}
-              </ReactTooltip>
-            </b>
-          </td>
-          <td>{row.hprefix.toString()}</td>
-        </tr>
-        <tr>
-          <td>
-            <b>
-              <div data-tip data-for={'mprefix_exp'}>
-                {<span>{'Hijacked Prefix'}</span>}
-              </div>
-              <ReactTooltip html={true} id={'mprefix_exp'}>
-                {
-                  'The configured IPv4/IPv6 prefix that matched the hijacked prefix.'
-                }
-              </ReactTooltip>
-            </b>
-          </td>
-          <td>{row.mprefix.toString()}</td>
-        </tr>
-        <tr>
-          <td>
-            <b>
-              <div data-tip data-for={'type_exp'}>
-                {<span>{'Type'}</span>}
-              </div>
-              <ReactTooltip html={true} id={'type_exp'}>
-                {
-                  'The type of the hijack in 4 dimensions: prefix|path|data plane|policy.'
-                }
-              </ReactTooltip>
-            </b>
-          </td>
-          <td>{row.type}</td>
-        </tr>
-        <tr>
-          <td>
-            <b>
-              <div data-tip data-for={'rpki_exp'}>
-                {<span>{'RPKI'}</span>}
-              </div>
-              <ReactTooltip html={true} id={'rpki_exp'}>
-                {'The RPKI status of the hijacked prefix.'}
-              </ReactTooltip>
-            </b>
-          </td>
-          <td>{row.rpki.toString()}</td>
-        </tr>
-        <tr>
-          <td>
-            <b>
-              <div data-tip data-for={'seen_exp'}>
-                {<span>{'# Peers Seen'}</span>}
-              </div>
-              <ReactTooltip html={true} id={'seen_exp'}>
-                {
-                  'Number of peers/monitors (i.e., ASNs)<br>that have seen hijack updates.'
-                }
-              </ReactTooltip>
-            </b>
-          </td>
-          <td>{row.peers.toString()}</td>
-        </tr>
-        <tr>
-          <td>
-            <b>
-              <div data-tip data-for={'inf_exp'}>
-                {<span>{'# ASes Infected'}</span>}
-              </div>
-              <ReactTooltip html={true} id={'inf_exp'}>
-                {
-                  'Number of infected ASes that seem to<br>route traffic towards the hijacker AS.<br>Note that this is an experimental field.'
-                }
-              </ReactTooltip>
-            </b>
-          </td>
-          <td>{row.ASes.toString()}</td>
-        </tr>
-        <tr>
-          <td>
-            <b>
-              <div data-tip data-for={'ack_exp'}>
-                {<span>{'ACK'}</span>}
-              </div>
-              <ReactTooltip html={true} id={'ack_exp'}>
-                {
-                  'Whether the user has acknowledged/confirmed the hijack as a true positive.<br>If the resolve|mitigate buttons are pressed this<br>is automatically set to True (default value: False).'
-                }
-              </ReactTooltip>
-            </b>
-          </td>
-          <td>
-            {row.resolved || row.under_mitigation ? (
-              <img alt="" src="./handled.png" />
-            ) : (
-              <img alt="" src="./unhadled.png" />
-            )}
-          </td>
-        </tr>
+        <thead></thead>
+        <tbody>
+          <tr>
+            <td>
+              <b>
+                <div data-tip data-for={'hprefix_exp'}>
+                  {<span>{'Hijacked Prefix'}</span>}
+                </div>
+                <ReactTooltip html={true} id={'hprefix_exp'}>
+                  {'The IPv4/IPv6 prefix that was hijacked.'}
+                </ReactTooltip>
+              </b>
+            </td>
+            <td>{row.hprefix.toString()}</td>
+          </tr>
+          <tr>
+            <td>
+              <b>
+                <div data-tip data-for={'mprefix_exp'}>
+                  {<span>{'Hijacked Prefix'}</span>}
+                </div>
+                <ReactTooltip html={true} id={'mprefix_exp'}>
+                  {
+                    'The configured IPv4/IPv6 prefix that matched the hijacked prefix.'
+                  }
+                </ReactTooltip>
+              </b>
+            </td>
+            <td>{row.mprefix.toString()}</td>
+          </tr>
+          <tr>
+            <td>
+              <b>
+                <div data-tip data-for={'type_exp'}>
+                  {<span>{'Type'}</span>}
+                </div>
+                <ReactTooltip html={true} id={'type_exp'}>
+                  {
+                    'The type of the hijack in 4 dimensions: prefix|path|data plane|policy.'
+                  }
+                </ReactTooltip>
+              </b>
+            </td>
+            <td>{row.type}</td>
+          </tr>
+          <tr>
+            <td>
+              <b>
+                <div data-tip data-for={'rpki_exp'}>
+                  {<span>{'RPKI'}</span>}
+                </div>
+                <ReactTooltip html={true} id={'rpki_exp'}>
+                  {'The RPKI status of the hijacked prefix.'}
+                </ReactTooltip>
+              </b>
+            </td>
+            <td>{row.rpki.toString()}</td>
+          </tr>
+          <tr>
+            <td>
+              <b>
+                <div data-tip data-for={'seen_exp'}>
+                  {<span>{'# Peers Seen'}</span>}
+                </div>
+                <ReactTooltip html={true} id={'seen_exp'}>
+                  {
+                    'Number of peers/monitors (i.e., ASNs)<br>that have seen hijack updates.'
+                  }
+                </ReactTooltip>
+              </b>
+            </td>
+            <td>{row.peers.toString()}</td>
+          </tr>
+          <tr>
+            <td>
+              <b>
+                <div data-tip data-for={'inf_exp'}>
+                  {<span>{'# ASes Infected'}</span>}
+                </div>
+                <ReactTooltip html={true} id={'inf_exp'}>
+                  {
+                    'Number of infected ASes that seem to<br>route traffic towards the hijacker AS.<br>Note that this is an experimental field.'
+                  }
+                </ReactTooltip>
+              </b>
+            </td>
+            <td>{row.ASes.toString()}</td>
+          </tr>
+          <tr>
+            <td>
+              <b>
+                <div data-tip data-for={'ack_exp'}>
+                  {<span>{'ACK'}</span>}
+                </div>
+                <ReactTooltip html={true} id={'ack_exp'}>
+                  {
+                    'Whether the user has acknowledged/confirmed the hijack as a true positive.<br>If the resolve|mitigate buttons are pressed this<br>is automatically set to True (default value: False).'
+                  }
+                </ReactTooltip>
+              </b>
+            </td>
+            <td>
+              {row.resolved || row.under_mitigation ? (
+                <img alt="" src="./handled.png" />
+              ) : (
+                <img alt="" src="./unhadled.png" />
+              )}
+            </td>
+          </tr>
+        </tbody>
       </table>
     );
   },
