+++ apps/artemis-web/pages/hijack.tsx
@@ -18,7 +18,13 @@ import AuthHOC from '../components/401-hoc/401-hoc';
 import BGPTableComponent from '../components/bgp-table/bgp-table';
 import { fetchASNData } from '../utils/fetch-data';
 import { useGraphQl } from '../utils/hooks/use-graphql';
-import { extractHijackInfos, parseASNData } from '../utils/parsers';
+import {
+  extractDistinctTooltips,
+  extractHijackInfos,
+  extractHijackTooltips,
+  extractWithDrawnSeen,
+  parseASNData,
+} from '../utils/parsers';
 import { useStyles } from '../utils/styles';
 import { formatDate, fromEntries, genTooltip, shallMock } from '../utils/token';
 
@@ -144,91 +150,16 @@ const ViewHijackPage = (props) => {
   useEffect(() => {
     let isMounted = true;
     (async function setStateFn() {
-      const ASN_int_origin: number = hijack.hijack_as;
-      const ASN_int_peers: number = hijack.peers_seen;
+      const { tooltipsWithdrawn, tooltipsSeen } = await extractWithDrawnSeen(
+        withdrawn,
+        seen
+      );
 
-      const [name_origin, countries_origin, abuse_origin] =
-        ASN_int_origin && ASN_int_origin.toString() !== '-'
-          ? await fetchASNData(ASN_int_origin)
-          : ['', '', ''];
-      const [name_peers, countries_peers, abuse_peers] =
-        ASN_int_peers && ASN_int_peers.toString() !== '-'
-          ? await fetchASNData(ASN_int_peers)
-          : ['', '', ''];
-
-      const tooltipsWithdrawn = [];
-      const tooltipsSeen = [];
-      const waitData1 = [];
-      const waitData2 = [];
-
-      for (let i = 0; i < withdrawn.length; i++) {
-        waitData1.push(await fetchASNData(withdrawn[i]));
-        tooltipsWithdrawn.push(
-          parseASNData(
-            withdrawn[i],
-            waitData1[i][0],
-            waitData1[i][1],
-            waitData1[i][2]
-          )
-        );
-      }
-
-      for (let i = 0; i < seen.length; i++) {
-        waitData2.push(await fetchASNData(seen[i]));
-        tooltipsSeen.push(
-          parseASNData(
-            seen[i],
-            waitData2[i][0],
-            waitData2[i][1],
-            waitData2[i][2]
-          )
-        );
-      }
-
-      const tooltip1 =
-        ASN_int_origin && ASN_int_origin.toString() !== '-'
-          ? parseASNData(
-              ASN_int_origin,
-              name_origin,
-              countries_origin,
-              abuse_origin
-            )
-          : '';
-
-      const tooltip2 =
-        ASN_int_peers && ASN_int_peers.toString() !== '-'
-          ? parseASNData(
-              ASN_int_peers,
-              name_peers,
-              countries_peers,
-              abuse_peers
-            )
-          : '';
-
-      const tooltips = {};
-
-      for (let i = 0; i < asns.length; i++) {
-        const ASN_int: number | string =
-          asns[i] !== '-' ? parseInt(asns[i], 10) : '-';
-        const [name_origin, countries_origin, abuse_origin] =
-          ASN_int == '-' ? ['', '', ''] : await fetchASNData(ASN_int);
-
-        const tooltip =
-          ASN_int == '-'
-            ? ''
-            : parseASNData(
-                ASN_int,
-                name_origin,
-                countries_origin,
-                abuse_origin
-              );
-        tooltips[ASN_int] = tooltip;
-      }
       if (isMounted) {
-        setASNDistinctTitle(tooltips);
+        setASNDistinctTitle(await extractDistinctTooltips(asns));
         setASNWithdrawnTitle(tooltipsWithdrawn);
         setASNSeenTitle(tooltipsSeen);
-        setASNTitle([tooltip1, tooltip2]);
+        setASNTitle(await extractHijackTooltips(hijack));
       }
     })();
